refactor(ai-agent): split auxiliary_routes into routes/ with AgentInstance dispatch
